[C++] Demonstrate ack callback, recovery, headers provider, Arrow compression inline in examples
