+++ cpp/examples/arrow/README.md
@@ -116,13 +116,8 @@ zerobus::ArrowStream stream = sdk.create_arrow_stream(
 - `Zstd` — higher compression ratio, more CPU per batch.
 - `NoCompression` (the default) — no compression.
 
-A complete runnable version is in [`compression.cpp`](compression.cpp) — it is
-`arrow_ingest.cpp` with `ipc_compression` set to `Zstd`. Build and run it the
-same way:
-
-```bash
-./build/examples/arrow_compression
-```
+`arrow_ingest.cpp` shows this as a commented-out option on
+`ArrowStreamOptions`; uncomment it to enable compression.
 
 ## Adapting for Your Custom Table
