+++ cpp/examples/json/single.cpp
@@ -8,6 +8,10 @@
 // server round-trip per record and collapse throughput. For high volume, prefer
 // the batch API in batch.cpp.
 //
+// It also demonstrates recovery: if the stream fails terminally, the records it
+// never acknowledged can be recovered via get_unacked_records() and re-ingested
+// on a fresh stream.
+//
 // Configuration — every connection setting is read from the environment, so no
 // value is ever baked into source. Export these before running (see
 // ../README.md for what each one is and the full copy-pasteable block):
@@ -26,6 +30,7 @@
 #include <cstdlib>
 #include <iostream>
 #include <string>
+#include <vector>
 
 #include "zerobus/zerobus.hpp"
 
@@ -64,6 +69,16 @@ std::string make_order_json(int id, const std::string& customer,
          ", \"updated_at\": " + std::to_string(ts) + "}";
 }
 
+zerobus::Stream open_stream(zerobus::Sdk& sdk, const std::string& table_name,
+                            const std::string& client_id,
+                            const std::string& client_secret) {
+  zerobus::TableProperties props;
+  props.table_name = table_name;
+  zerobus::StreamOptions options;
+  options.record_type = zerobus::RecordType::Json;
+  return sdk.create_stream(props, client_id, client_secret, options);
+}
+
 }  // namespace
 
 int main() {
@@ -85,14 +100,8 @@ int main() {
     // 2. Open a JSON stream. record_type must be Json to match the payloads,
     //    and descriptor_proto is left empty (no schema needed for JSON — the
     //    server maps each record's fields onto the table's columns by name).
-    zerobus::TableProperties props;
-    props.table_name = table_name;
-
-    zerobus::StreamOptions options;
-    options.record_type = zerobus::RecordType::Json;
-
     zerobus::Stream stream =
-        sdk.create_stream(props, client_id, client_secret, options);
+        open_stream(sdk, table_name, client_id, client_secret);
 
     const std::int64_t now = now_micros();
 
@@ -116,17 +125,39 @@ int main() {
         std::to_string(now) + ", \"updated_at\": " + std::to_string(now) + "}");
     std::cout << "Record 3 queued with offset ID: " << offset << "\n";
 
-    // 4. Flush once at the end: block until every queued record is durably
-    //    acknowledged by the server. This is the right place to wait — not
-    //    after each individual ingest above.
-    stream.flush();
-    std::cout << "All records acknowledged.\n";
+    // 4. Flush once, then close — both at a controlled point. close() surfaces
+    //    any final error by throwing; ~Stream() would swallow it.
+    //
+    //    Guard these with a nested try/catch to demonstrate recovery. The SDK
+    //    recovers transparently from transient disconnects; only a TERMINAL
+    //    failure throws here, and a failed close() keeps the handle alive so
+    //    get_unacked_records() can hand back whatever was never acknowledged.
+    //    (After a *successful* close the handle is freed, so that call would
+    //    throw instead — recovery belongs on the failure path only.)
+    try {
+      stream.flush();
+      stream.close();
+      std::cout << "All records acknowledged. Stream closed successfully.\n";
+    } catch (const zerobus::ZerobusException& e) {
+      std::cerr << "Stream failed: " << e.what() << "\n";
 
-    // 5. Close at a controlled point rather than leaving it to the destructor.
-    //    close() surfaces any final error by throwing; ~Stream() would swallow
-    //    it.
-    stream.close();
-    std::cout << "Stream closed successfully.\n";
+      std::vector<zerobus::UnackedRecord> unacked =
+          stream.get_unacked_records();
+      std::cout << "Recovering " << unacked.size()
+                << " unacknowledged records on a fresh stream.\n";
+
+      if (!unacked.empty()) {
+        zerobus::Stream retry =
+            open_stream(sdk, table_name, client_id, client_secret);
+        // Re-ingest, then flush once — the same loop-then-flush pattern.
+        for (const zerobus::UnackedRecord& record : unacked) {
+          retry.ingest_json_record(record.as_string());
+        }
+        retry.flush();
+        retry.close();
+        std::cout << "Recovered records re-ingested and acknowledged.\n";
+      }
+    }
   } catch (const zerobus::ZerobusException& e) {
     std::cerr << "Zerobus error: " << e.what()
               << " (retryable=" << (e.is_retryable() ? "true" : "false")
